+++ Jump_n_Bump/src/de/oetting/bumpingbunnies/usecases/game/businesslogic/CollisionHandling.java
@@ -30,11 +30,11 @@ public class CollisionHandling {
 		GameObject nextStepX = player.simulateNextStepX();
 		if (collisionDetection.collides(nextStepX, object)) {
 			if (player.movementX() > 0) {
-				int diffX = object.minX() - player.maxX();
+				int diffX = (int) (object.minX() - player.maxX());
 				player.setMovementX(diffX);
 				player.setAccelerationX(0);
 			} else if (player.movementX() < 0) {
-				int diffX = object.maxX() - player.minX();
+				int diffX = (int) (object.maxX() - player.minX());
 				player.setMovementX(diffX);
 				player.setAccelerationX(0);
 			}
@@ -46,11 +46,11 @@ public class CollisionHandling {
 		GameObject nextStepY = player.simulateNextStepY();
 		if (collisionDetection.collides(nextStepY, object)) {
 			if (player.movementY() > 0) {
-				int diffY = object.minY() - player.maxY();
+				int diffY = (int) (object.minY() - player.maxY());
 				player.setMovementY(diffY);
 				player.setAccelerationY(0);
 			} else if (player.movementY() < 0) {
-				int diffY = object.maxY() - player.minY();
+				int diffY = (int) (object.maxY() - player.minY());
 				player.setMovementY(diffY);
 				player.setAccelerationY(0);
 			}
