+++ Jump_n_Bump/src/de/oetting/bumpingbunnies/usecases/game/graphics/Animation.java
@@ -7,7 +7,7 @@ public interface Animation {
 	public abstract void updateGraphics(CanvasDelegate canvas, int width,
 			int height);
 
-	public abstract void draw(CanvasDelegate canvas, int left, int top,
+	public abstract void draw(CanvasDelegate canvas, long left, long top,
 			Paint paint);
 
 }
