+++ Jump_n_Bump/src/de/oetting/bumpingbunnies/usecases/game/model/IcyWall.java
@@ -2,7 +2,7 @@ package de.oetting.bumpingbunnies.usecases.game.model;
 
 public class IcyWall extends Wall {
 
-	public IcyWall(int id, int minX, int minY, int maxX, int maxY) {
+	public IcyWall(int id, long minX, long minY, long maxX, long maxY) {
 		super(id, minX, minY, maxX, maxY, 0xff8888ff);
 	}
